fix(notifications): avoid settings module cycle
