+++ .changeset/neat-notification-context.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-notifications': patch
+---
+
+Updated internal imports to avoid circular module dependencies in notification settings.
