+++ plugins/notifications/src/components/UserNotificationSettingsCard/UserNotificationSettingsCard.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { createContext, useState, useContext, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { ErrorPanel, InfoCard, Progress } from '@backstage/core-components';
 import { useNotificationsApi } from '../../hooks';
 import { NotificationSettings } from '@backstage/plugin-notifications-common';
@@ -23,60 +23,7 @@ import { useApi } from '@backstage/core-plugin-api';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 import { notificationsTranslationRef } from '../../translation';
 import { UserNotificationSettingsPanel } from './UserNotificationSettingsPanel';
-import { capitalize } from 'lodash';
-
-type FormatContextType = {
-  formatOriginName: (id: string) => string;
-  formatTopicName: (id: string) => string;
-};
-
-const NotificationFormatContext = createContext<FormatContextType | undefined>(
-  undefined,
-);
-
-export const useNotificationFormat = () => {
-  const { t } = useTranslationRef(notificationsTranslationRef);
-  const context = useContext(NotificationFormatContext);
-  if (!context) throw new Error(t('settings.errors.useNotificationFormat'));
-  return context;
-};
-
-type Props = {
-  children: React.ReactNode;
-  originMap: Record<string, string> | undefined;
-  topicMap: Record<string, string> | undefined;
-};
-
-export const NotificationFormatProvider = ({
-  children,
-  originMap,
-  topicMap,
-}: Props) => {
-  const formatName = (
-    id: string,
-    nameMap: Record<string, string> | undefined,
-  ) => {
-    if (nameMap && id in nameMap) {
-      return nameMap[id];
-    }
-    return capitalize(id.replaceAll(/[-_:]/g, ' '));
-  };
-
-  const formatOriginName = (originId: string) => {
-    return formatName(originId, originMap);
-  };
-
-  const formatTopicName = (topicId: string) => {
-    return formatName(topicId, topicMap);
-  };
-  return (
-    <NotificationFormatContext.Provider
-      value={{ formatOriginName, formatTopicName }}
-    >
-      {children}
-    </NotificationFormatContext.Provider>
-  );
-};
+import { NotificationFormatProvider } from './NotificationFormatContext';
 
 /** @public */
 export const UserNotificationSettingsCard = (props: {
